# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a DEIP server API.",
    "version": "1.0.0",
    "title": "DEIP",
    "contact": {
      "email": "yahor.tsaryk@deip.world"
    }
  },
...
  "paths": {
...
    "/api/v2/teams": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "Find teams",
        "description": "Return an array of teams",
        "operationId": "getTeams",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Team"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v2/teams/listing": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "Find teams listing",
        "description": "Return an array of teams",
        "operationId": "getTeamsListing",
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Team"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v2/teams/portal/{portalId}": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "Find teams",
        "description": "Return an array of portal teams",
        "operationId": "getTeamsByPortal",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "portalId",
            "in": "path",
            "description": "portalId of teams to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Team"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/team/logo/{teamId}": {
      "get": {
        "tags": [
          "team"
        ],
        "summary": "Find team logo",
        "description": "Return team logo",
        "operationId": "getTeamLogo",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "teamId of team",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Team"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Team not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "Team": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "portalId": {
          "type": "string"
        },
        "attributes": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "value": {},
              "attributeId": {
                "type": "string"
              }
            }
          }
        },
        "creator": {
          "type": "string"
        },
        "balances": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "account": {
                "type": "string"
              },
              "assetId": {
                "type": "string"
              },
              "symbol": {
                "type": "string"
              },
              "precision": {
                "type": "number"
              },
              "amount": {
                "type": "string"
              }
            }
          }
        },
        "members": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "metadataHash": {
          "type": "string"
        }
      }
    },
...
  }
}